🐛Renew foreign device registration even in case of error or no response
The behavior of register/unregister was a bit a confusing:

- You would call register and if everything was perfect then
  registration renewals would be done automatically.
- Except when the BBMD would sometimes report an error in which case
  registration renewal 'loop' would stop.
- Or when the BBMD would not respond to subsequent registration request
  (e.g. unreliable network) in which case the renewal loop would stop
  and the BIPForeign instance will find itself in an inconsistent state
  (see[*] bellow).

To solve this, make BIPForeign a RecurringTask so that renewals will be
sent continuously (unless `unregister` is explicitly called). Whenever
an ACK is received, (re)start a timer that will set the
`registrationStatus` to unregistered if no subsequent ACK comes from the
BBMD in 30s (J.5.2.3) + `bbmdTimeToLive`.

[*]: issue registration request, BBMD ACKs, issue registration request
(renewal), BBMD does not respond (e.g. datagram dropped while on route
to the BBMD). In this situation the registrationStatus remains
'registered' and no further registration renewal requests are issued.